+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 ##CHANGE LOG
 ---
 
+##1.0.1 - 02/15/2016 - [Levon Becker](mailto:devops@bonusbits.com)
+
+* Add nodeinfo scripts
+* Updated Rakefile and Gemfile
+* Updated Readme
+* Changed TravisCI rake task to "ci"
+* Added install package list as attribute so can be overroad.
+* Added draft Chefspec
+* Added coveralls yaml
+
 ##1.0.0 - 06/29/2015 - [Levon Becker](mailto:devops@bonusbits.com)
 
 * Initial Release